Add files via upload
The Weather Analysis and Prediction Web Application project demonstrates the integration of Python's data science capabilities with web development to create a user-friendly tool for weather analysis and forecasting. The inclusion of a pickle file enhances the project's efficiency by storing the trained predictive model, while the web interface makes weather insights and predictions accessible to a wider audience. This project showcases the potential of combining different aspects of programming to address real-world challenges.